ChipChat_Project/requirements.txt
google-genai
python-dotenv
rich>=13.0.0
